agents/platform/platform-agent/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "platform-agent"
version = "1.0.0"
description = "Platform Agent — وكيل المنصة لمنظومة نافع"
authors = [{ name = "Nafic", email = "hello@nafic.com" }]
requires-python = ">=3.12"
dependencies = [
    # Graph / AI
    "langgraph>=0.2.0",
    "langchain-core>=0.2.0",
    "anthropic>=0.25.0",

    # HTTP
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "httpx>=0.27.0",

    # Database
    "psycopg2-binary>=2.9.9",

    # Redis
    "redis[hiredis]>=5.0.4",

    # Email
    "resend>=0.7.2",

    # Validation
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.1",

    # Utilities
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "respx>=0.21.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["platform_agent*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]
